moved CompilerPass to Bundle Extension PROCERGS#638
