+++ src/LoginCidadao/OpenIDBundle/Resources/config/services.yml
@@ -51,7 +51,7 @@ services:
             - { name: jms_serializer.event_listener, event: serializer.post_serialize, method: onPostSerialize }
 
     oauth2.response_types.token:
-        class: OAuth2\ResponseType\AccessToken
+        class: LoginCidadao\OpenIDBundle\ResponseType\AccessToken
         arguments:
             - "@oauth2.storage.access_token"
             - "@oauth2.storage.refresh_token"
